+++ echoframework/src/main/kotlin/org/echo/mobile/framework/facade/internal/TransactionsFacadeImpl.kt
@@ -46,14 +46,10 @@ class TransactionsFacadeImpl(
             AuthorityType.ACTIVE
         )
 
-        val memoPrivateKey = memoKey(fromAccount.name, password)
-        val memo = generateMemo(memoPrivateKey, fromAccount, toAccount, message)
-
         val transfer = TransferOperationBuilder()
             .setFrom(fromAccount)
             .setTo(toAccount)
             .setAmount(AssetAmount(UnsignedLong.valueOf(amount.toLong()), Asset(asset)))
-            .setMemo(memo)
             .build()
 
         val transaction = configureTransaction(transfer, privateKey, asset, feeAsset)
@@ -76,13 +72,11 @@ class TransactionsFacadeImpl(
         checkOwnerAccount(wif, fromAccount)
 
         val privateKey = cryptoCoreComponent.decodeFromWif(wif)
-        val memo = generateMemo(privateKey, fromAccount, toAccount, message)
 
         val transfer = TransferOperationBuilder()
             .setFrom(fromAccount)
             .setTo(toAccount)
             .setAmount(AssetAmount(UnsignedLong.valueOf(amount.toLong()), Asset(asset)))
-            .setMemo(memo)
             .build()
 
         val transaction = configureTransaction(transfer, privateKey, asset, feeAsset)