+++ echoframework/src/main/kotlin/org/echo/mobile/framework/model/EthAddress.kt
@@ -8,7 +8,8 @@ import com.google.gson.annotations.SerializedName
  * @author Dmitriy Bushuev
  */
 data class EthAddress(
-    @SerializedName("acc_id") val accountId: String,
+    val id: String,
+    @SerializedName("account") val accountId: String,
     @SerializedName("eth_addr") val address: String,
     @SerializedName("is_approved") val isApproved: Boolean,
     val approves: List<String>