+++ echoframework/src/main/kotlin/org/echo/mobile/framework/model/operations/WithdrawEthereumOperation.kt
@@ -90,7 +90,7 @@ class WithdrawEthereumOperation constructor(
     }
 
     companion object {
-        private const val ACCOUNT_ID_KEY = "acc_id"
+        private const val ACCOUNT_ID_KEY = "account"
         private const val ETH_ADDRESS_KEY = "eth_addr"
         private const val VALUE_KEY = "value"
     }