+++ echoframework/src/test/kotlin/org/echo/mobile/framework/EchoFrameworkTest.kt
@@ -16,7 +16,9 @@ import org.echo.mobile.framework.model.GlobalProperties
 import org.echo.mobile.framework.model.HistoryResponse
 import org.echo.mobile.framework.model.Log
 import org.echo.mobile.framework.model.Price
+import org.echo.mobile.framework.model.TransactionResult
 import org.echo.mobile.framework.model.contract.ContractBalance
+import org.echo.mobile.framework.model.contract.ContractFee
 import org.echo.mobile.framework.model.contract.ContractInfo
 import org.echo.mobile.framework.model.contract.ContractResult
 import org.echo.mobile.framework.model.contract.ContractStruct
@@ -69,13 +71,13 @@ class EchoFrameworkTest {
         )
     }
 
-    private val legalContractId = "1.14.1"
-    private val legalTokenId = "1.14.0"
-    private val accountId = "1.2.22"
+    private val legalContractId = "1.14.0"
+    private val legalTokenId = "1.14.1"
+    private val accountId = "1.2.16"
     private val login = "dima"
     private val password = "dima"
     private val wif = "5J3UbadSyzzcQQ7HEfTr2brhJJpHhx3NsMzrvgzfysBesutNRCm"
-    private val secondAccountId = "1.2.23"
+    private val secondAccountId = "1.2.19"
     private val secondLogin = "daria"
     private val secondPassword = "daria"
     private val legalAssetId = "1.3.0"
@@ -325,7 +327,7 @@ class EchoFrameworkTest {
 
         if (connect(framework) == false) Assert.fail("Connection error")
 
-        framework.getAccount("dima", futureAccount.completeCallback())
+        framework.getAccount("daria", futureAccount.completeCallback())
 
         val account = futureAccount.get()
         assertTrue(account != null)
@@ -398,10 +400,10 @@ class EchoFrameworkTest {
     }
 
     @Test
-    fun accountHistoryByIdTest() = getAccountHistory(accountId)
+    fun accountHistoryByIdTest() = getAccountHistory("1.2.8")
 
     @Test
-    fun accountHistoryByNameTest() = getAccountHistory("vsharaev")
+    fun accountHistoryByNameTest() = getAccountHistory("daria")
 
     private fun getAccountHistory(nameOrId: String) {
         val framework = initFramework()
@@ -414,7 +416,7 @@ class EchoFrameworkTest {
             nameOrId,
             "1.10.0",
             "1.10.0",
-            20,
+            100,
             futureAccountHistory.completeCallback()
         )
 
@@ -536,14 +538,40 @@ class EchoFrameworkTest {
     fun getEthereumTest() {
         val framework = initFramework()
 
-        val futureChangePassword = FutureTask<List<EthAddress>>()
+        val futureEthereum = FutureTask<EthAddress>()
 
         if (connect(framework) == false) Assert.fail("Connection error")
 
-        framework.getEthereumAddresses(
-            "daria1",
-            object : Callback<List<EthAddress>> {
-                override fun onSuccess(result: List<EthAddress>) {
+        framework.getEthereumAddress(
+            "daria",
+            object : Callback<EthAddress> {
+                override fun onSuccess(result: EthAddress) {
+                    futureEthereum.setComplete(result)
+                }
+
+                override fun onError(error: LocalException) {
+                    error.printStackTrace()
+                    futureEthereum.setComplete(error)
+                }
+
+            })
+
+        assertNotNull(futureEthereum.get() ?: false)
+    }
+
+    @Test
+    fun withdrawEthereumTest() {
+        val framework = initFramework()
+
+        val futureChangePassword = FutureTask<Boolean>()
+        val futureResult = FutureTask<TransactionResult>()
+
+        if (connect(framework) == false) Assert.fail("Connection error")
+
+        framework.ethWithdraw(
+            "daria","daria","0x46Ba2677a1c982B329A81f60Cf90fBA2E8CA9fA8","1", "1.3.0",
+            object : Callback<Boolean> {
+                override fun onSuccess(result: Boolean) {
                     futureChangePassword.setComplete(result)
                 }
 
@@ -552,37 +580,11 @@ class EchoFrameworkTest {
                     futureChangePassword.setComplete(error)
                 }
 
-            })
+            }, futureResult.completeCallback())
 
-        assertNotNull(futureChangePassword.get() ?: false)
+        val result = futureResult.get()
+        assertNotNull(result ?: false)
     }
-//
-//    @Test
-//    fun withdrawEthereumTest() {
-//        val framework = initFramework()
-//
-//        val futureChangePassword = FutureTask<Boolean>()
-//        val futureResult = FutureTask<TransactionResult>()
-//
-//        if (connect(framework) == false) Assert.fail("Connection error")
-//
-//        framework.ethWithdraw(
-//            "vsharaev","vsharaev","d28224c2cfa050bff674160274e0d41742c7ec43","1", "1.3.0",
-//            object : Callback<Boolean> {
-//                override fun onSuccess(result: Boolean) {
-//                    futureChangePassword.setComplete(result)
-//                }
-//
-//                override fun onError(error: LocalException) {
-//                    error.printStackTrace()
-//                    futureChangePassword.setComplete(error)
-//                }
-//
-//            }, futureResult.completeCallback())
-//
-//        val result = futureResult.get()
-//        assertNotNull(result ?: false)
-//    }
 //
 //    @Test
 //    fun withdrawEthereumWithWifTest() {
@@ -1056,7 +1058,7 @@ class EchoFrameworkTest {
 
         if (connect(framework) == false) Assert.fail("Connection error")
 
-        val futureFee = FutureTask<String>()
+        val futureFee = FutureTask<ContractFee>()
 
         framework.getFeeForContractOperation(
             userNameOrId = accountId,
@@ -1078,7 +1080,7 @@ class EchoFrameworkTest {
 
         if (connect(framework) == false) Assert.fail("Connection error")
 
-        val futureFee = FutureTask<String>()
+        val futureFee = FutureTask<ContractFee>()
 
         framework.getFeeForContractOperation(
             userNameOrId = "1.2.16",
@@ -1099,7 +1101,7 @@ class EchoFrameworkTest {
 
         if (connect(framework) == false) Assert.fail("Connection error")
 
-        val futureFee = FutureTask<String>()
+        val futureFee = FutureTask<ContractFee>()
 
         framework.getFeeForContractOperation(
             userNameOrId = login,
@@ -1228,7 +1230,7 @@ class EchoFrameworkTest {
             "daria",
             "daria",
             asset = "1.3.2",
-            amount = "100",
+            amount = "1",
             destinationIdOrName = "daria",
             message = "Do it",
             callback = futureIssue.completeCallback()
@@ -1416,8 +1418,8 @@ class EchoFrameworkTest {
         framework.callContractWithWif(
             "daria",
             "5J9YnfSUx6GnweorDEswRNAFcBzsZrQoJLkfqKLzXwBdRvjmoz1",
-            assetId = "1.3.0",
-            feeAsset = "1.3.0",
+            assetId = legalAssetId,
+            feeAsset = legalAssetId,
             contractId = legalContractId,
             methodName = "logTest",
             methodParams = listOf(),
@@ -1669,7 +1671,7 @@ class EchoFrameworkTest {
         val future = FutureTask<ContractResult>()
 
         framework.getContractResult(
-            historyId = "1.15.1",
+            historyId = "1.15.0",
             callback = future.completeCallback()
         )